pyproject.toml
[build-system]
requires = ['setuptools>=68', 'wheel']
build-backend = 'setuptools.backends.legacy:build'

[project]
name = 'vulnnet-env'
version = '1.0.0'
requires-python = '>=3.12'
dependencies = ['fastapi', 'uvicorn', 'pydantic>=2', 'openenv-core', 'openai']

[project.scripts]
server = "server.app:main"
